Increase max_execution_time to 9999
The php ini-setting "max_execution_time" seems to be seriously broken in combination with Imagick.
When converting animated GIFs a value of 240 is breached within a few seconds.
Also the value 0 (no limit) is breached immediately.
Since we have a CGI timeout limit in apache it's safe to basically disable the php timeout.