Don't write apt file "update-success-stamp"
This file is part of our puppet module, but writing it here is not technically needed and it's probably better to not rely on it.
The only upside I can think of is that the initial installation of a host that DOESN'T have puppet yet will be a bit faster, but that happens rarely.